Issue Fixed
If the target's/player's inventory is full, the item will drop on the ground. If it can be placed in the inventory, it goes into the inventory.

Used the isFull method and changed it to do what I needed it to do. Commented out an unused method I saw (for cleanliness sake).

Should be good. Let me know if theres anything else that needs to be fixed.